+++ src/nvim/testing.h
@@ -1,0 +1,10 @@
+#ifndef NVIM_TESTING_H
+#define NVIM_TESTING_H
+
+#include "nvim/eval/typval.h"
+#include "nvim/eval/funcs.h"
+
+#ifdef INCLUDE_GENERATED_DECLARATIONS
+# include "testing.h.generated.h"
+#endif
+#endif  // NVIM_TESTING_H
